requirements-aws.txt
# AWS Lambda requirements
# Include all existing requirements plus AWS-specific packages

# Existing requirements from requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
cryptography==41.0.7
python-multipart==0.0.6
structlog==23.2.0
httpx==0.25.2
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# AWS-specific packages
mangum==0.17.0          # ASGI adapter for AWS Lambda
boto3==1.34.0           # AWS SDK
botocore==1.34.0        # AWS core library

# OpenAI for LLM services
openai==1.3.7

# Additional dependencies for production
pydantic-settings==2.1.0
python-dotenv==1.0.0    # For local development

# Monitoring and logging
aws-lambda-powertools==2.25.0
